requirements.txt

flask
pytesseract
Pillow
pdf2image
